TASK #138 Move SourceParser to interface
