+++ src/NullDev/Nemesis/services.yml
@@ -48,4 +48,4 @@ services:
 
   NullDev\Skeleton\Uuid\Handler\CreateUuidClassHandler: ~
 
-  NullDev\Nemesis\SourceParser: ~
+  NullDev\Nemesis\PhpParserSourceParser: ~
